+++ Block/Config.php
@@ -35,42 +35,60 @@ class Config extends \Magento\Framework\View\Element\Template
     private $amazonConfig;
 
     /**
-     * Config constructor.
+     * @var \Amazon\Pay\Model\Subscription\SubscriptionManager
+     */
+    private $subscriptionManager;
+
+    /**
+     * Config constructor
+     *
      * @param \Magento\Framework\View\Element\Template\Context $context
      * @param \Amazon\Pay\Helper\Data $amazonHelper
      * @param \Amazon\Pay\Model\AmazonConfig $amazonConfig
+     * @param \Amazon\Pay\Model\Subscription\SubscriptionManager $subscriptionManager
      */
     public function __construct(
         \Magento\Framework\View\Element\Template\Context $context,
         \Amazon\Pay\Helper\Data $amazonHelper,
-        \Amazon\Pay\Model\AmazonConfig $amazonConfig
+        \Amazon\Pay\Model\AmazonConfig $amazonConfig,
+        \Amazon\Pay\Model\Subscription\SubscriptionManager $subscriptionManager
     ) {
         parent::__construct($context);
         $this->amazonHelper = $amazonHelper;
         $this->amazonConfig = $amazonConfig;
+        $this->subscriptionManager = $subscriptionManager;
     }
 
     /**
+     * Package module configuration values for button rendering
+     *
      * @return array
      */
     public function getConfig()
     {
         $config = [
-            'region'                   => $this->amazonConfig->getRegion(),
-            'code'                     => \Amazon\Pay\Gateway\Config\Config::CODE,
-            'is_method_available'      => $this->amazonConfig->isPayButtonAvailableAsPaymentMethod(),
-            'is_pay_only'              => $this->amazonHelper->isPayOnly(),
+            'region'                    => $this->amazonConfig->getRegion(),
+            'code'                      => \Amazon\Pay\Gateway\Config\Config::CODE,
+            'vault_code'                => \Amazon\Pay\Gateway\Config\Config::VAULT_CODE,
+            'is_method_available'       => $this->amazonConfig->isPayButtonAvailableAsPaymentMethod(),
+            'is_pay_only'               => $this->amazonHelper->isPayOnly(),
             'is_lwa_enabled'            => $this->isLwaEnabled(),
             'is_guest_checkout_enabled' => $this->amazonConfig->isGuestCheckoutEnabled(),
             'has_restricted_products'   => $this->amazonHelper->hasRestrictedProducts(),
-            'is_multicurrency_enabled'     => $this->amazonConfig->multiCurrencyEnabled(),
-            'acceptance_mark' => $this->amazonConfig->getAcceptanceMark()
+            'is_multicurrency_enabled'  => $this->amazonConfig->multiCurrencyEnabled(),
+            'acceptance_mark'           => $this->amazonConfig->getAcceptanceMark()
         ];
+
+        if ($subscriptionLabel = $this->subscriptionManager->getSubscriptionLabel()) {
+            $config['subscription_label'] = $subscriptionLabel;
+        }
 
         return $config;
     }
 
     /**
+     * Convert config values to JSON object
+     *
      * @return string
      */
     public function getJsonConfig()
@@ -79,6 +97,8 @@ class Config extends \Magento\Framework\View\Element\Template
     }
 
     /**
+     * Return true if module is enabled
+     *
      * @return bool
      */
     public function isEnabled()
@@ -87,6 +107,8 @@ class Config extends \Magento\Framework\View\Element\Template
     }
 
     /**
+     * Return true if Amazon Sign in is enabled
+     *
      * @return bool
      */
     public function isLwaEnabled()