+++ Model/Exception/OrderFailureException.php
@@ -1,0 +1,7 @@
+<?php
+namespace Amazon\Pay\Model\Exception;
+
+class OrderFailureException extends \Exception
+{
+
+}