+++ Plugin/PaymentTransactionIdUpdate.php
@@ -22,9 +22,11 @@ use Magento\Sales\Model\Order\Payment\Transaction\ManagerInterface;
 class PaymentTransactionIdUpdate
 {
     /**
+     * Set special transaction ID if AP transaction is voided
+     *
      * @param ManagerInterface $subject
      * @param OrderPaymentInterface $payment
-     * @param $type
+     * @param string $type
      * @param bool $transactionBasedOn
      * @return array
      * @SuppressWarnings(PHPMD.UnusedFormalParameter)